Forward some extension methods which is dependent to UnityEngine.AssetBundleModule
