+++ src/UniTask/Assets/Plugins/UniTask/Runtime/Modules.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 796ffa0906465b446ada92dd199f0092
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
