+++ src/UniTask/Assets/Plugins/UniTask/Runtime/Modules/AssetBundle.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c011199bfe2467b4f8a8a1c01c83d1e7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
