+++ src/UniTask/Assets/Plugins/UniTask/Runtime/Modules/AssetBundle/AssetBundleAsyncExtensions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 31079e7527b74f947b26df8b2626f36e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
