+++ src/UniTask/Assets/Plugins/UniTask/Runtime/Modules/AssetBundle/UniTask.Modules.AssetBundle.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c0a8c2dd3ebbbd04b860af0504f5d7b8
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
